refactor: extract common math function to mathCommon
