+++ packages/server/src/config/config.js
@@ -70,7 +70,7 @@ const markdown = {
     emoji: !isFalse(MARKDOWN_EMOJI),
     sub: !isFalse(MARKDOWN_SUB),
     sup: !isFalse(MARKDOWN_SUP),
-    // tex is false | 'katex' | other default MathJax.
+    // tex is false | 'katex' | 'mathjax' (default and fallback).
     tex: !isFalse(MARKDOWN_TEX) ? MARKDOWN_TEX : false,
   },
 };
